+++ gatsby-config.js
@@ -53,6 +53,13 @@ module.exports = {
         pathToConfigModule: `src/utils/typography.js`,
       },
     },
+    {
+      resolve: `gatsby-plugin-postcss-sass`,
+      options: {
+        postCssPlugins: [require('autoprefixer')()],
+        precision: 8, // SASS default: 5
+      },
+    },
     'gatsby-plugin-react-helmet',
     'gatsby-plugin-sass',
     {