+++ src/layouts/links.js
@@ -1,0 +1,25 @@
+const links = [
+  {
+    href:
+      'https://fonts.googleapis.com/css?family=Roboto:300,300i,400,400i,500,500i,700,700i,900',
+    rel: 'stylesheet',
+  },
+  {
+    href:
+      'https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.css',
+    rel: 'stylesheet',
+    type: 'text/css',
+  },
+  {
+    href: 'https://fonts.googleapis.com/css?family=Nunito:400,700"',
+    rel: 'stylesheet',
+    type: 'text/css',
+  },
+  {
+    href:
+      'https://fonts.googleapis.com/css?family=Roboto+Mono:400,400i,700,700i',
+    rel: 'stylesheet',
+  },
+]
+
+export default links